Implemented Render System
Cleaned up app class, implemented a Renderer class, as well as a simple_render_system class. The Renderer class handles creation of the swapchain as well as creating the commnandBuffers. The simple_render_system class handles the creation of the pipeline and pipeline layout. It also handles the rendering of game objects.